refactor(quant): make scale-storage ABI states unrepresentable
QuantWeight now owns its format/aux invariant for its whole lifetime:
both fields are private, and the single validation in
check_aux_matches_format runs at construction (new) and at every retag
(with_format). A weight can no longer be flipped to a format whose
kernels read an external scale buffer it does not carry - the plain
field assignment that produced the q8 QKV SIGTRAP is now
unrepresentable. The supposed 65 production mutation sites collapsed to
47 writes in one test fixture; production never mutated format.

Fabricated scale buffers are gone from the fixtures. The q8 QKV fused
test and the new hybrid Q8_0 smoke test build genuine Q8_0
representation (int8 rows + per-32-block f32 external scales, matching
q8_qkv_proj's `Ws + row * blocks` indexing) instead of retagging
scale-less Q4_K bytes. full_pipeline_seq1_produces_nonzero fed Q4_0
weights an ExternalScales buffer that was actually input-quantization
scales; Q4_0 packs its f16 scale inline, so it now passes QuantAux::None
- and still passes, confirming the full pipeline genuinely supports
Q4_0 QKV via q4_matvec.

Operand-local dispatch (O-projection keyed on wo, QKV on wq) exposed
these test-path bugs rather than reinstating fake buffers: the hybrid
path refuses Q4_0 attention loudly (no route can serve it), with a
should_panic test replacing the smoke test that only "passed" by
handing q8_matvec Q4_0 blocks as int8 rows. ENV_TEST_LOCK acquisition
is poison-tolerant everywhere so the intentional-panic tests cannot
cascade.

The synthetic Metal decode suite is split from one 3424-line file into
tests/metal_decode_synthetic/ (main + common fixture + 10 topic
modules, largest 713 lines) with its 77-test coverage unchanged.
