+++ crates/larql-compute-metal/src/decode/encode_attn.rs
@@ -73,7 +73,6 @@ pub(super) struct AttnBufs<'a> {
 pub(super) struct AttnDims {
     pub hidden: usize,
     pub layer_q_dim: usize,
-    pub uses_kquant: bool,
     /// True iff the FFN side will run Q4_K family (selects the fused
     /// `residual_norm_store` path that mirrors the FFN's input dtype).
     pub ffn_uses_kquant: bool,
@@ -95,7 +94,6 @@ impl MetalBackend {
         let AttnDims {
             hidden,
             layer_q_dim,
-            uses_kquant,
             ffn_uses_kquant,
         } = dims;
         let hidden_val = hidden as u32;
@@ -481,73 +479,101 @@ impl MetalBackend {
         }
 
         // ── Step 5a: O projection ──
-        if uses_kquant {
-            use crate::stages::quant_matvec::Pipelines;
-            let pipes = Pipelines {
-                q4kf_proj: Some(&self.attention.q4kf_proj_pipeline.state),
-                q4k_matvec_fallback: &self.quant.q4k_matvec_pipeline,
-                q6k_matvec: &self.quant.q6k_matvec_pipeline,
-                q4_matvec: &self.q4.matvec,
-                q4k_matmul: None,
-            };
-            crate::stages::o_proj::encode(
-                enc,
-                &pipes,
-                &self.quant.q8_quant_pipeline,
-                layer.wo.format,
-                bufs.wo,
-                bufs.attn_out_buf,
-                0,
-                bufs.o_q8_scratch,
-                0,
-                bufs.o_q8s_scratch,
-                0,
-                bufs.o_out_buf,
-                0,
-                layer_q_dim,
-                hidden,
-            );
-        } else {
+        //
+        // Dispatched on `wo`'s own format, not `wq`'s. Selecting an
+        // operand's ABI from a *different* operand's representation is
+        // what produced the defect this replaced: `uses_kquant` reads
+        // `wq`, so a Q4_0 `wo` fell to the Q8-only branch below and was
+        // handed to `q8_matvec` as though its 18-byte packed blocks were
+        // int8 rows, alongside a weight-scale buffer Q4_0 has no source
+        // for. That was survivable only because every caller fabricated
+        // an empty scale buffer; removing the fabrication exposed it.
+        match layer.wo.format() {
+            // `o_proj::encode` understands all four packed formats — its
+            // own doc covers "Q4_0 / Q8_0: quantise attn_in -> Q8 int8 +
+            // per-32 f16 scale" as well as the k-quant families.
+            larql_compute::QuantFormat::Q4_0
+            | larql_compute::QuantFormat::Q4_K
+            | larql_compute::QuantFormat::Q4_KF
+            | larql_compute::QuantFormat::Q6_K => {
+                use crate::stages::quant_matvec::Pipelines;
+                let pipes = Pipelines {
+                    q4kf_proj: Some(&self.attention.q4kf_proj_pipeline.state),
+                    q4k_matvec_fallback: &self.quant.q4k_matvec_pipeline,
+                    q6k_matvec: &self.quant.q6k_matvec_pipeline,
+                    q4_matvec: &self.q4.matvec,
+                    q4k_matmul: None,
+                };
+                crate::stages::o_proj::encode(
+                    enc,
+                    &pipes,
+                    &self.quant.q8_quant_pipeline,
+                    layer.wo.format(),
+                    bufs.wo,
+                    bufs.attn_out_buf,
+                    0,
+                    bufs.o_q8_scratch,
+                    0,
+                    bufs.o_q8s_scratch,
+                    0,
+                    bufs.o_out_buf,
+                    0,
+                    layer_q_dim,
+                    hidden,
+                );
+            }
             // Q8 legacy path: decode-specific `q8_matvec` shader (not in
-            // stages::quant_matvec which uses `q4_matvec` for Q4_0/Q8_0 with
-            // a different buffer layout). Inline.
-            let dim_val = layer_q_dim as u32;
-            let blocks = (layer_q_dim / LEGACY_BLOCK_ELEMS) as u32;
-            enc.set_compute_pipeline_state(&self.quant.q8_quant_pipeline);
-            enc.set_buffer(0, Some(bufs.attn_out_buf), 0);
-            enc.set_buffer(1, Some(bufs.o_q8_scratch), 0);
-            enc.set_buffer(2, Some(bufs.o_q8s_scratch), 0);
-            enc.set_bytes(3, 4, &dim_val as *const u32 as *const std::ffi::c_void);
-            enc.dispatch_threads(
-                MTLSize::new(blocks as u64, 1, 1),
-                MTLSize::new(
-                    crate::kernels::DISPATCH_TG_MAX_THREADS.min(blocks as u64),
-                    1,
-                    1,
-                ),
-            );
+            // stages::quant_matvec, which uses `q4_matvec` with a
+            // different buffer layout). Q8_0 only — it is the one format
+            // that supplies the external weight scales this shader reads
+            // at buffer index 2.
+            larql_compute::QuantFormat::Q8_0 => {
+                let dim_val = layer_q_dim as u32;
+                let blocks = (layer_q_dim / LEGACY_BLOCK_ELEMS) as u32;
+                enc.set_compute_pipeline_state(&self.quant.q8_quant_pipeline);
+                enc.set_buffer(0, Some(bufs.attn_out_buf), 0);
+                enc.set_buffer(1, Some(bufs.o_q8_scratch), 0);
+                enc.set_buffer(2, Some(bufs.o_q8s_scratch), 0);
+                enc.set_bytes(3, 4, &dim_val as *const u32 as *const std::ffi::c_void);
+                enc.dispatch_threads(
+                    MTLSize::new(blocks as u64, 1, 1),
+                    MTLSize::new(
+                        crate::kernels::DISPATCH_TG_MAX_THREADS.min(blocks as u64),
+                        1,
+                        1,
+                    ),
+                );
 
-            let o_rows = hidden as u32;
-            let o_k = layer_q_dim as u32;
-            enc.set_compute_pipeline_state(&self.quant.q8_matvec_pipeline.state);
-            enc.set_buffer(0, Some(bufs.wo), 0);
-            enc.set_buffer(1, Some(bufs.o_q8_scratch), 0);
-            enc.set_buffer(
-                2,
-                Some(
-                    bufs.wo_scales
-                        .expect("legacy scale path requires an external-scale format"),
-                ),
-                0,
-            );
-            enc.set_buffer(3, Some(bufs.o_q8s_scratch), 0);
-            enc.set_buffer(4, Some(bufs.o_out_buf), 0);
-            enc.set_bytes(5, 4, &o_rows as *const u32 as *const std::ffi::c_void);
-            enc.set_bytes(6, 4, &o_k as *const u32 as *const std::ffi::c_void);
-            enc.dispatch_thread_groups(
-                MTLSize::new((hidden as u64).div_ceil(8), 1, 1),
-                MTLSize::new(256, 1, 1),
-            );
+                let o_rows = hidden as u32;
+                let o_k = layer_q_dim as u32;
+                enc.set_compute_pipeline_state(&self.quant.q8_matvec_pipeline.state);
+                enc.set_buffer(0, Some(bufs.wo), 0);
+                enc.set_buffer(1, Some(bufs.o_q8_scratch), 0);
+                enc.set_buffer(
+                    2,
+                    Some(
+                        bufs.wo_scales
+                            .expect("legacy scale path requires an external-scale format"),
+                    ),
+                    0,
+                );
+                enc.set_buffer(3, Some(bufs.o_q8s_scratch), 0);
+                enc.set_buffer(4, Some(bufs.o_out_buf), 0);
+                enc.set_bytes(5, 4, &o_rows as *const u32 as *const std::ffi::c_void);
+                enc.set_bytes(6, 4, &o_k as *const u32 as *const std::ffi::c_void);
+                enc.dispatch_thread_groups(
+                    MTLSize::new((hidden as u64).div_ceil(8), 1, 1),
+                    MTLSize::new(256, 1, 1),
+                );
+            }
+            // Loud rather than silent. The previous predicate routed
+            // anything non-k-quant here, so an unsupported format was
+            // reinterpreted as Q8 and produced plausible numbers instead
+            // of an error.
+            format => panic!(
+                "O projection has no dispatch for {format:?}; supported: \
+                 Q4_0, Q4_K, Q4_KF, Q6_K (o_proj::encode) and Q8_0 (legacy q8_matvec)"
+            ),
         }
 
         // ── Step 5b: Residual + post-attn norm + ffn-input norm ──
