+++ crates/larql-compute-metal/src/decode_hybrid.rs
@@ -48,7 +48,33 @@ impl MetalBackend {
         } else {
             layer_head_dim
         };
-        let uses_kquant = layer.wq.format.is_kquant_family();
+        // Operand-local ABI selection. A single `uses_kquant` boolean read
+        // from `wq` used to route QKV, the O projection and their scale
+        // bindings alike — so one operand's representation decided how a
+        // different operand was interpreted. Q4_0 fell to the Q8 branches
+        // and had its 18-byte packed blocks handed to `q8_qkv_proj` /
+        // `q8_matvec` as though they were int8 rows, which only survived
+        // because callers fabricated an empty scale buffer.
+        //
+        // Q4_0 attention is not supported by any route here: the Q8
+        // kernels need int8 weights plus external scales, and Q4_0 has
+        // neither. Production cannot construct it either —
+        // `attn_str_to_format` admits only Q4_K and Q6_K. Refuse loudly
+        // rather than reinterpret the bytes.
+        assert_ne!(
+            layer.wq.format(),
+            larql_compute::QuantFormat::Q4_0,
+            "Q4_0 attention weights are not supported: the k-quant route does not \
+             handle Q4_0 QKV and the Q8 route needs external scales Q4_0 has no \
+             source for"
+        );
+        assert_ne!(
+            layer.wo.format(),
+            larql_compute::QuantFormat::Q4_0,
+            "Q4_0 O-projection weights are not supported on the hybrid path"
+        );
+        let qkv_uses_kquant = layer.wq.format().is_kquant_family();
+        let o_uses_kquant = layer.wo.format().is_kquant_family();
         let layer_q_dim = layer_num_q_heads * layer_head_dim;
         let window_size = layer.sliding_window as u32;
 
@@ -89,7 +115,7 @@ impl MetalBackend {
 
         let enc_a = cmd.new_compute_command_encoder();
 
-        if uses_kquant {
+        if qkv_uses_kquant {
             use crate::ops::full_pipeline::encode_rms_norm;
             let norm_f32_buf = self.bufs.output((hidden * 4) as u64);
             let total_rows = (q_dim + kv_dim + kv_dim) as u32;
@@ -116,7 +142,7 @@ impl MetalBackend {
             // the other's pipeline silently drops 75 % of QKV rows.
             // Same dispatch-geometry-mismatch class as the q4_matvec_v4
             // ROADMAP ship-log entry.
-            let qkv_pipeline = if layer.wq.format == larql_compute::QuantFormat::Q4_KF {
+            let qkv_pipeline = if layer.wq.format() == larql_compute::QuantFormat::Q4_KF {
                 &self.attention.q4kf_qkv_proj_pipeline
             } else {
                 &self.attention.q4k_qkv_proj_pipeline
@@ -333,14 +359,14 @@ impl MetalBackend {
 
         let enc_c = cmd.new_compute_command_encoder();
 
-        // O projection
-        if uses_kquant {
+        // O projection — dispatched on `wo`'s own format.
+        if o_uses_kquant {
             let o_rows = hidden as u32;
             let o_k = layer_q_dim as u32;
             let o_out = self.bufs.output((hidden * 4) as u64);
-            let o_pipeline = if layer.wo.format == larql_compute::QuantFormat::Q4_KF {
+            let o_pipeline = if layer.wo.format() == larql_compute::QuantFormat::Q4_KF {
                 &self.attention.q4kf_proj_pipeline
-            } else if layer.wo.format == larql_compute::QuantFormat::Q6_K {
+            } else if layer.wo.format() == larql_compute::QuantFormat::Q6_K {
                 &self.quant.q6k_matvec_pipeline
             } else {
                 &self.quant.q4k_matvec_pipeline
