+++ crates/larql-compute-metal/tests/test_metal_shaders.rs
@@ -1397,32 +1397,35 @@ fn full_pipeline_seq1_produces_nonzero() {
     let wk_data = quantize_q4_0(&vec![0.01f32; kv_dim * hidden]);
     let wv_data = quantize_q4_0(&vec![0.01f32; kv_dim * hidden]);
     let wo_data = quantize_q4_0(&vec![0.01f32; hidden * q_dim]);
-    let (_q8_x_q, q8_s_q) = q4::quantize_to_q8(&vec![0.01f32; hidden]);
 
     let norm = vec![1.0f32; hidden];
     let x: Vec<f32> = (0..hidden).map(|i| (i as f32 * 0.01).sin()).collect();
 
+    // Q4_0 packs its f16 scale inside each 18-byte block, so no external
+    // scale array exists. This fixture used to fabricate one out of
+    // *input*-quantization scales — the wrong object entirely — which
+    // `QuantWeight::new` now refuses.
     let layer = larql_compute::FullPipelineLayer {
         attn_sinks: None,
         wq: larql_compute::QuantWeight::new(
             larql_compute::QuantFormat::Q4_0,
             &wq_data,
-            larql_compute::QuantAux::ExternalScales(&q8_s_q),
+            larql_compute::QuantAux::None,
         ),
         wk: larql_compute::QuantWeight::new(
             larql_compute::QuantFormat::Q4_0,
             &wk_data,
-            larql_compute::QuantAux::ExternalScales(&q8_s_q),
+            larql_compute::QuantAux::None,
         ),
         wv: larql_compute::QuantWeight::new(
             larql_compute::QuantFormat::Q4_0,
             &wv_data,
-            larql_compute::QuantAux::ExternalScales(&q8_s_q),
+            larql_compute::QuantAux::None,
         ),
         wo: larql_compute::QuantWeight::new(
             larql_compute::QuantFormat::Q4_0,
             &wo_data,
-            larql_compute::QuantAux::ExternalScales(&q8_s_q),
+            larql_compute::QuantAux::None,
         ),
         gate: larql_compute::QuantWeight::new(
             larql_compute::QuantFormat::Q4_0,
